rename validators util into model-helpers folder
