+++ ui/lib/core/addon/components/kv-object-editor.js
@@ -10,7 +10,12 @@ import { assert } from '@ember/debug';
 import { action } from '@ember/object';
 import { guidFor } from '@ember/object/internals';
 import KVObject from 'vault/lib/kv-object';
-import { hasWhitespace, isNonString, NON_STRING_WARNING, WHITESPACE_WARNING } from 'vault/utils/validators';
+import {
+  hasWhitespace,
+  isNonString,
+  NON_STRING_WARNING,
+  WHITESPACE_WARNING,
+} from 'vault/utils/model-helpers/validators';
 
 /**
  * @module KvObjectEditor
